:bug: fix : Fix merge conflicts

+++ backend/airlingo/src/main/java/com/ssafy/airlingo/domain/user/entity/User.java
@@ -155,7 +155,7 @@ public class User extends BaseTimeEntity {
 		this.userComplain += 1;
 	}
 
-	public void renewRatingAndStudyCount(float rating) {
+	public void renewRatingAndStudyCount(double rating){
 		userTotalRating += rating;
 		userStudyCount += 1;
 		userRating = userTotalRating / userStudyCount;
